add docker

+++ src/main/resources/db/migration/V001__Create_table.sql
@@ -6,36 +6,36 @@ drop table if exists make;
 
 create table make
 (
-	id   		SERIAL PRIMARY KEY,
+	id   		BIGSERIAL PRIMARY KEY,
 	name		varchar(255) UNIQUE not null
 );
 
 create table model
 (
-	id   		SERIAL PRIMARY KEY,
-	make_ref	INT REFERENCES make (id) on delete cascade not null,
+	id   		BIGSERIAL PRIMARY KEY,
+	make_ref	BIGINT REFERENCES make (id) on delete cascade not null,
 	name		varchar(255) UNIQUE not null
 );
 
 create table category
 (
-	id   		SERIAL PRIMARY KEY,
+	id   		BIGSERIAL PRIMARY KEY,
 	name		varchar(255) UNIQUE not null
 );
 
 create table car
 (
-    id   		SERIAL PRIMARY KEY,
+    id   		BIGSERIAL PRIMARY KEY,
     object_id	varchar(255) UNIQUE not null,
     year 		INT not null,
-    make_ref	INT REFERENCES make (id) on delete cascade not null,
-    model_ref	INT REFERENCES model (id) on delete cascade not null
+    make_ref	BIGINT REFERENCES make (id) on delete cascade not null,
+    model_ref	BIGINT REFERENCES model (id) on delete cascade not null
 );
 
 create table car_category
 (
-	car_id		INT REFERENCES car (id) on delete cascade not null,
-	category_id	INT REFERENCES category (id) on delete cascade not null,
+	car_id		BIGINT REFERENCES car (id) on delete cascade not null,
+	category_id	BIGINT REFERENCES category (id) on delete cascade not null,
 	UNIQUE		(car_id, category_id)
 );
